Create base Join
