+++ quixstreams/dataframe/joins/join_asof.py
@@ -1,14 +1,10 @@
 import typing
-from datetime import timedelta
-from typing import Any, Callable, Optional, Union
 
 from quixstreams.context import message_context
-from quixstreams.dataframe.utils import ensure_milliseconds
 from quixstreams.models.topics.manager import TopicManager
 from quixstreams.state.rocksdb.timestamped import TimestampedPartitionTransaction
 
-from .base import JoinHow, JoinHow_choices, OnOverlap, OnOverlap_choices
-from .utils import keep_left_merger, keep_right_merger, raise_merger
+from .base import Join
 
 if typing.TYPE_CHECKING:
     from quixstreams.dataframe.dataframe import StreamingDataFrame
@@ -19,39 +15,7 @@ __all__ = ("AsOfJoin",)
 DISCARDED = object()
 
 
-class AsOfJoin:
-    def __init__(
-        self,
-        how: JoinHow,
-        on_merge: Union[OnOverlap, Callable[[Any, Any], Any]],
-        grace_ms: Union[int, timedelta],
-        store_name: Optional[str] = None,
-    ):
-        if how not in JoinHow_choices:
-            raise ValueError(
-                f'Invalid "how" value: {how}. '
-                f"Valid choices are: {', '.join(JoinHow_choices)}."
-            )
-        self._how = how
-
-        if callable(on_merge):
-            self._merger = on_merge
-        elif on_merge == "keep-left":
-            self._merger = keep_left_merger
-        elif on_merge == "keep-right":
-            self._merger = keep_right_merger
-        elif on_merge == "raise":
-            self._merger = raise_merger
-        else:
-            raise ValueError(
-                f'Invalid "on_merge" value: {on_merge}. '
-                f"Provide either one of {', '.join(OnOverlap_choices)} or "
-                f"a callable to merge records manually."
-            )
-
-        self._grace_ms = ensure_milliseconds(grace_ms)
-        self._store_name = store_name or "join"
-
+class AsOfJoin(Join):
     def join(
         self,
         left: "StreamingDataFrame",
